RF: Move Namespace into types module
