+++ tools/schemacode/bidsschematools/types/__init__.py
@@ -1,0 +1,1 @@
+from .namespace import Namespace
